refactor(CvCard): moving sub-components into their own directories, adjusting data type to accept normalised data
