+++ src/components/CvCard/CvCard.tsx
@@ -4,11 +4,11 @@ import { View } from 'react-native'
 import { Colors } from '../../styles'
 import { WithChildren } from '../../types/react.types'
 import { isNotNil } from '../../utils/ramda.utils'
+import EditButton from '../EditButton/EditButton'
 import ListCardHeader from '../ListCardHeader'
 import Optional from '../Optional'
 import Text, { TextAlign } from '../Typography'
 import styles from './CvCard.styles'
-import EditButton from './EditButton'
 
 type Props = WithChildren<{
   title: string
