+++ src/components/CvCard/CvCardListBody/CvCardListBody.styles.ts
