+++ src/components/CvCard/CvCardListBody/index.ts
@@ -1,0 +1,3 @@
+import CvCardListBody from './CvCardListBody'
+
+export default CvCardListBody
